added routing and re-organized the folder structures
